requirements.txt
pynacl==1.5.0
cbor2==5.4.6
argon2-cffi==21.3.0
